third_party/blink/web_tests/fast/forms/time-multiple-fields/time-multiple-fields-change-layout-by-value-expected.txt

Change multiple fields time input UI layout by value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS widthOfStep60 < widthWithMillisecond is true
PASS widthOfStep60 == widthWithSecond is true
PASS successfullyParsed is true

TEST COMPLETE
